Broke ModulesForInitialRegistration into its own file.
