+++ ui/src/app/pages/access/ducc-card.tsx
@@ -5,12 +5,9 @@ import { AccessModule, Profile } from 'generated/fetch';
 import { FlexColumn, FlexRow } from 'app/components/flex';
 import { WithSpinnerOverlayProps } from 'app/components/with-spinner-overlay';
 
-import {
-  DARPageMode,
-  ModulesForInitialRegistration,
-  styles,
-} from './data-access-requirements';
+import { DARPageMode, styles } from './data-access-requirements';
 import { ModulesForAnnualRenewal } from './modules-for-annual-renewal';
+import { ModulesForInitialRegistration } from './modules-for-initial-registration';
 
 const duccModule = AccessModule.DATAUSERCODEOFCONDUCT;
 export const DuccCard = (props: {
